Исправления после ревью

+++ src/main/java/ru/yandex/practicum/filmorate/storage/db/mapper/MpaRowMapper.java
@@ -11,7 +11,7 @@ public class MpaRowMapper implements RowMapper<Mpa> {
     @Override
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
-        mpa.setId(rs.getInt("id"));
+        mpa.setId(rs.getLong("id"));
         mpa.setName(rs.getString("name"));
         return mpa;
     }
